[OG-51] Home refactor
